clean up: rename lib file
