+++ pcmdi_metrics/enso/lib/__init__.py
@@ -1,1 +1,1 @@
-from .PMPdriver_lib import AddParserArgument, tree, metrics_to_json, find_realm, get_file, CLIVAR_LargeEnsemble_Variables  # noqa
+from .enso_lib import AddParserArgument, tree, metrics_to_json, find_realm, get_file, CLIVAR_LargeEnsemble_Variables  # noqa
